+++ platform/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v1.1.0] platform
 ### Releases
 - [NPM v1.1.0] https://www.npmjs.com/package/@codat/platform/v/1.1.0 - platform
+
+## 2023-09-21 00:03:51
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.88.1 (2.122.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.1.1] platform
+### Releases
+- [NPM v1.1.1] https://www.npmjs.com/package/@codat/platform/v/1.1.1 - platform